Changed price field to float

+++ moontracker/views/home/views.py
@@ -2,7 +2,8 @@
 from flask import request, render_template, flash, Blueprint
 from flask_wtf import RecaptchaField, Recaptcha
 from flask_login import current_user
-from wtforms import Form, StringField, IntegerField, SelectField, validators
+from wtforms import Form, FloatField, StringField, IntegerField, SelectField
+from wtforms import validators
 import json
 
 from moontracker.assets import supported_assets, assets, market_apis
@@ -93,7 +94,7 @@ class AlertForm(Form):
                 '^[0-9]+$', message="Input characters must be numeric")])
     asset = SelectField(
         'Coin', choices=assets)
-    target_price = IntegerField('Target Price', [validators.optional()])
+    target_price = FloatField('Target Price', [validators.optional()])
     less_more = SelectField(
         '', choices=[(1, 'above'), (0, 'below')], coerce=int)
     recaptcha = RecaptchaField(
@@ -152,7 +153,7 @@ class MarketsForm(Form):
                           validators=product_validators)
 
     target_price_validators = [validators.InputRequired()]
-    target_price = IntegerField('Target Price',
+    target_price = FloatField('Target Price',
                                 validators=target_price_validators)
 
     less_more_choices = [(1, 'above'), (0, 'below')]
@@ -184,7 +185,7 @@ class ProductsForm(Form):
                          default='', validators=market_validators)
 
     target_price_validators = [validators.InputRequired()]
-    target_price = IntegerField('Target Price',
+    target_price = FloatField('Target Price',
                                 validators=target_price_validators)
 
     less_more_choices = [(1, 'above'), (0, 'below')]
